add base host

+++ host/manger.go
@@ -19,32 +19,45 @@ import (
 )
 
 type manager struct {
-	BaseDir      string
-	BaseHostFile string
-	Hosts        map[string]*Host
-	Groups       map[string][]*Host
+	BaseDir  string
+	BaseHost *Host
+	Hosts    map[string]*Host
+	Groups   map[string][]*Host
 }
 
 var Manager *manager
 
 func init() {
+	// init manager
 	Manager = &manager{
-		BaseDir:      conf.BaseDir,
-		BaseHostFile: conf.BaseHostFile,
-		Hosts:        map[string]*Host{},
-		Groups:       map[string][]*Host{},
+		BaseDir: conf.BaseDir,
+		BaseHost: &Host{
+			Name:     conf.BaseHostFileName,
+			FileName: conf.BaseHostFileName,
+			Path:     conf.BaseHostFile,
+			Groups:   nil,
+		},
+		Hosts:  map[string]*Host{},
+		Groups: map[string][]*Host{},
 	}
+
+	// create base dir
 	if _, err := os.Stat(Manager.BaseDir); os.IsNotExist(err) {
 		if err := os.Mkdir(Manager.BaseDir, os.ModePerm); err != nil {
-			panic("can not create dir " + Manager.BaseDir)
+			display.Panic("can not create dir "+Manager.BaseDir, err)
 		}
-		fmt.Println("create host dir", Manager.BaseDir)
 	}
-	if _, err := os.Stat(Manager.BaseHostFile); os.IsNotExist(err) {
-		if _, err := os.Create(Manager.BaseHostFile); err != nil {
-			panic("can not create base host file")
+
+	// create base host file
+	if _, err := os.Stat(Manager.BaseHost.Path); os.IsNotExist(err) {
+		var content bytes.Buffer
+		content.WriteString("127.0.0.1 localhost")
+		content.WriteByte(NewLine)
+		content.WriteString("::1 localhost")
+		content.WriteByte(NewLine)
+		if err := os.WriteFile(Manager.BaseHost.Path, content.Bytes(), 0644); err != nil {
+			display.Panic("can not create base host file", err)
 		}
-		fmt.Println("create base host file", Manager.BaseHostFile)
 	}
 	Manager.LoadHosts()
 }
@@ -54,8 +67,9 @@ func (m *manager) LoadHosts() {
 	if err != nil {
 		display.ErrExit(fmt.Errorf("failed to load gohost dir"))
 	}
+	// load host files
 	for _, file := range files {
-		// skip dir and .* files
+		// skip dir and files started with '.'
 		if file.IsDir() || strings.HasPrefix(file.Name(), ".") {
 			continue
 		}
@@ -70,7 +84,6 @@ func (m *manager) PrintGroups() {
 		fmt.Println("no host group")
 		return
 	}
-
 	header := []string{"Group", "Hosts"}
 	data := make([][]string, 0, len(m.Groups))
 	for group, hosts := range m.Groups {
@@ -145,9 +158,11 @@ func (m *manager) AddGroup(hostName string, groups []string) {
 }
 
 func (m *manager) CreateNewHost(name string, groups []string) {
+	if name == m.BaseHost.Name {
+		display.ErrExit(fmt.Errorf("host file '%s' already exists\n", name))
+	}
 	if _, exist := m.Hosts[name]; exist {
-		display.ErrExit(fmt.Errorf("host file '%s' is existed\n", name))
-		return
+		display.ErrExit(fmt.Errorf("host file '%s' already exists\n", name))
 	}
 	filePath := m.fullFilePath(m.hostName(name, groups))
 	err := editor.OpenByVim(filePath)
@@ -172,6 +187,9 @@ func (m *manager) DeleteHostsByNames(hostNames []string) {
 }
 
 func (m *manager) ChangeHostName(hostName string, newHostName string) {
+	if hostName == m.BaseHost.Name || newHostName == m.BaseHost.Name {
+		display.ErrExit(fmt.Errorf("can not change base host file name"))
+	}
 	h := m.mustHost(hostName)
 	_newHostName := m.hostName(newHostName, h.Groups)
 	if err := os.Rename(h.Path, m.fullFilePath(_newHostName)); err != nil {
@@ -187,28 +205,41 @@ func (m *manager) EditHostFile(hostName string) {
 	}
 }
 
-func (m *manager) ApplyGroup(group string) {
+func (m *manager) ApplyGroup(group string, simulate bool) {
 	hosts, exist := m.Groups[group]
 	if !exist {
 		display.ErrExit(fmt.Errorf("not found group '%s'", group))
 		return
 	}
+	hosts = append(hosts, m.BaseHost)
 	combinedHostContent := m.combineHosts(hosts, "# Auto generated from group "+group)
+
+	// just print
+	if simulate {
+		fmt.Println(string(combinedHostContent))
+		return
+	}
+
+	// write to tmp file
 	combinedHost := m.fullFilePath(conf.TmpCombinedHost)
 	if err := ioutil.WriteFile(combinedHost, combinedHostContent, 0664); err != nil {
 		display.ErrExit(err)
 	}
-	if err := os.Rename(combinedHost, sysHost); err != nil {
+
+	// replace system host
+	if err := os.Rename(combinedHost, SysHost); err != nil {
 		display.ErrExit(err)
 	}
 	fmt.Printf("applied group '%s' to system host:\n", group)
+
+	// display system host
 	m.PrintSysHost(10)
 }
 
 func (m *manager) PrintSysHost(max int) {
-	host, err := os.Open(sysHost)
+	host, err := os.Open(SysHost)
 	if err != nil {
-		panic(err)
+		display.Panic("can not read system host file", err)
 	}
 	defer host.Close()
 	scanner := bufio.NewScanner(host)
@@ -226,6 +257,10 @@ func (m *manager) PrintSysHost(max int) {
 }
 
 func (m *manager) host(hostName string) (*Host, bool) {
+	if hostName == m.BaseHost.Name {
+		return m.BaseHost, true
+	}
+
 	host, exist := m.Hosts[hostName]
 	if !exist {
 		display.ErrExit(fmt.Errorf("host file '%s' is not existed\n", hostName))
@@ -235,6 +270,10 @@ func (m *manager) host(hostName string) (*Host, bool) {
 }
 
 func (m *manager) mustHost(hostName string) *Host {
+	if hostName == m.BaseHost.Name {
+		return m.BaseHost
+	}
+
 	host, exist := m.Hosts[hostName]
 	if !exist {
 		display.ErrExit(fmt.Errorf("host file '%s' is not existed\n", hostName))
@@ -293,7 +332,7 @@ func (m *manager) combineHosts(hosts []*Host, head string) []byte {
 	for _, host := range hosts {
 		file, err := os.Open(host.Path)
 		if err != nil {
-			panic(err)
+			display.Panic("can not combine host", err)
 		}
 		scanner := bufio.NewScanner(file)
 		b.WriteString("# Host section from " + host.Name + "\n")
